requirements.txt
cs50
feedparser
Flask
python-twitter
weather-api